requirements.txt
dacite==1.6.0
PyYAML==6.0.1
psycopg2-binary<=3.0.0; sys_platform == 'darwin'
psycopg2<=3.0.0; sys_platform == 'linux'
alembic==1.6.5
click==8.1.3
fastapi==0.115.0
uvicorn>=0.11.7
pydantic<3.0.0
python-multipart==0.0.5
py-consul==1.5.3

